Drop enum xtime_clock_types, update uses of boost::TIME_UTC
glibc newly defines a macro TIME_UTC, which collides with boost::TIME_UTC.
We can't avoid expanding that macro, but the value happens to be the same
as that of boost::TIME_UTC.  So drop enum xtime_clock_types.  Update boost
to use macro TIME_UTC instead of the scoped enum value.  External clients
will have to do the same.